Add files via upload
Merkle trees are used in bitcoin to summarize all the transactions in a block, produc‐
ing an overall digital fingerprint of the entire set of transactions, providing a very effi‐
cient process to verify whether a transaction is included in a block. A merkle tree is
constructed by recursively hashing pairs of nodes until there is only one hash, called
the root, or merkle root. The cryptographic hash algorithm used in bitcoin’s merkle
trees is SHA256 applied twice, also known as double-SHA256.